elastic: list page size and task view
